requirements.txt
fastapi==0.103.1
numpy==1.26.0
pillow==10.0.1
pre-commit==3.5.0
python-multipart==0.0.6
transformers==4.34.0
uvicorn[standard]==0.23.2
websockets==11.0.3
loguru==0.7.0
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-instrumentation-asgi==0.40b0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-instrumentation-logging==0.40b0
opentelemetry-exporter-jaeger==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-prometheus==1.12.0rc1
prometheus-client==0.17.1